Split out long-running tests
